requirements_prod.txt
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
xhtml2pdf==0.2.11
pypdf==3.15.4
python-decouple==3.8
psycopg2-binary